Added check for single view, multiple shortcodes
This checks whether the form is the same for the entry being shown and
the view. If not, it keeps going.

This only checks the form, which isn't perfect because different views
can use the same form. Not sure how else to do this, but for now it's
better than nothing.