Fix "Add View" shortcode generation
The shortcode JS needed to be manually updated for each new setting.

This new code reads through the settings and automatically generates
the embed code based on whether the settings are the default values or
have changed. If they have changed, then add a shortcode parameter.

Also, the name of the `gravityview_view_id` dropdown has been changed
to `gravityview_id` to match the name formatting of the rest of the
settings, where the name is `gravityview_{shortcode attribute name}`.