Add request caching class and functionality
Caches entry requests based on the form IDs that are being requested.
Stores results in transients.

This reduces the number of queries by hundreds when dealing with a
large request, and saves .3+ seconds.

When an entry is added or modified with the same form, transients are
flushed for entry queries that use that form.

This needs to be tested. @gedex - can you look into this and see if you
can find any issues?