Fix Gravity Forms database speed issue
Gravity Forms' GFCache function was thrashing the database, causing
double the amount of time for the field_value() method to run.

The reason is that the cache was checking against a field value stored
in a transient every time `GFFormsModel::get_lead_field_value()` is
called.

What we're doing here is telling the GFCache that it's already checked
the transient and the value is false, forcing it to just use the
non-cached data, which is actually faster.