The slug was messing it up.
I wanted to pass the `post_type=gravityview&page=gravityview_settings`
in the URL to keep it consistent with the single site settings URL, but
that messed things up. Turns out it wasn't necessary anyway.

So, update the slug and patch the CSS class and ids to have the
settings page look good again.

fixes #435