Improve speed of GVCommon::get_users()
*Major database resource and speed improvement*

Affects Search Bar and GravityView Search Widgets with "Submitted by"
fields, and Edit Entry screen with the Change Entry Creator.

Instead of fetching `all` fields, define a limited set of fields to
return. This way, they're all returned in a *single query* instead of
what was happening before: a main query, then another query for each
user returned.

* Before:  1.6 ms query + ( .3ms * 750 users ) = 226ms; 751 queries per
page load
* After:  1.3 ms query; 1 query per page load

Because of the speed improvements, I've increased the default # of
users returned to 2000.