Create a new [gv_entry_link] shortcode
This replaces the `[gv_delete_entry_link]` and `[gv_edit_entry_link]`
shortcodes, and also adds a `read` option that makes it easy to
directly link to an entry.

Accepts the same attributes as the previous shortcodes:

```
$action What type of link to generate. Options: `read`, `edit`, and
`delete`. Default: `read`
$view_id Define the ID for the View. If not set, use current View ID,
if exists.
$entry_id ID of the entry to edit. If undefined, uses the current entry
ID, if exists.
$post_id ID of the base post or page to use for an embedded View
$link_atts Pass anchor tag attributes (`target=_blank` to open Edit
Entry link in a new window, for example)
$return What should the shortcode return: link HTML (`html`) or the URL
(`url`). Default: `html`
$field_values Only used for `action="edit"`. Parameters to pass in to
the prefill data in Edit Entry form. Uses the same format as Gravity
Forms "Allow field to be populated dynamically" {@see
https://www.gravityhelp.com/documentation/article/allow-field-to-be-popu
lated-dynamically/ }
```