Added `gform_is_encrypted_field` filter before fetching entries
This prevents the database from being called for each entry. The only
reason the field would be encrypted is if passwords are being shown. If
that's the case, then we'd have a conflict that could be resolved by
adding a hook here: `gravityview_before_get_entries` and removing the
filter.