Convert bulk action delete to just a delete button
The bulk action thing didn't really make sense. There was always one
action. The reason I kept it was to prevent accidental deleting; that's
where having the Javascript `confirm()` comes in.

The problem with this is now accidental deleting when JS is disabled,
but since that still requires boxes to be checked, I'm not sure that's
really a problem.