Add support for password fields
So, by default the password fields would include two inputs ("Enter
Password" and "Confirm Password"). This code removes the second field
and in the GV admin re-labels it to be the parent field name.

Then on the front-end, we do some acrobatics to get the field from the
form and use that label.

Yes, this is kind of ridiculous.

Fixes #713