Add support for Admin multiselect fields
This one's for @naomicbush

I wanted to have `GravityView_FieldType_multiselect` extend the
`GravityView_FieldType_select` class and have it check the class name
to modify the output, but that file wasn't included yet, and it was all
more complicated that it was worth.

The multiselect wasn't saving properly, which turned out to be an issue
with our `array_merge_recursive()` replacement function. Thanks to the
fix provided here, it was sorted out:
https://github.com/reduxframework/redux-framework/issues/2631#issuecomme
nt-147772300