Edit_Entry_Render test examples
This is an initial stab at harnessing the renderer. It's one big hack,
mocking state via globals and superglobals, jumping through hoops lit on
fire while riding a unicile in a bear costume.

There are bound to be a crapton of edge-cases and use-cases, since forms
are so different, there are hundreds of GF_Fields out there and the
default form that comes with the test barely touches 10 of them (i.e.
not even all core ones).