Start returning \GV\View instances
...in the \GravityView_View_Data class

It used to return an array of sorts that is considered a "view"
structure. Now it returns a legitimage \GV\View instance that can still
be accessed as an array against those "old" keys. This ensures
back-compatibility.

Next - we switch out old array-based usage for the simple cases of "id"
and "view_id" around our core.