Don't show the . gv-list-view-content container div if no fields
One of the most common things we see at GV support is people with empty
zones that have a border drawn around them. That's not nice at all.

The template system needs to have a revamp sometime down the line, but
for now, this should be an easy win.

![Configuration example](http://share.gravityview.co/P1IN/2EwKfU9n+)

@rafaehlers Do you see any issues with this? The other way to do this
would be to _not change the DOM_ but to add CSS rules regarding hiding
the empty `<div>`. Which do you think is better?