Fix #851 missing "offset" attribute default
The [gravityview] shortcode uses `shortcode_atts` along with the
defaults, which "sanitizes" all incoming attributes, limiting keys to
only those found in the defaults.

Added the missing "offset" default. As an added bonus (or not), this
will show the offset setting in the Add View shortcode modal. The offset
setting is not shown in the View Configuration, though.