Add \GV\Source class
Since we have internal sources, for fields like custom content and
whatnot, we have to introduce a \GV\Source parent class, which \GV\Form
sources will inherit. Plus a \GV\Internal_Source class that knows how to
spew out values when needed.

This is groundwork for the `\GV\Field::get_value()` function, which is
ultimately required in the renderers...