Fix `\GV\View::by_id` dependence on global state
`get_post` relies on a the `$post` global to resolve a `$post_id` set to
NULL. Which tries to retrieve the current global `$post` object when
called through `\GV\View::by_id` with NULL given in turn.

Make sure we don't rely on the global `$post` when trying to retrieve a
`\GV\View` by a falsey ID (0, null, false, '', etc.)

Discovered as underlying problem in issue #870, and fixes it.