Ease up on the fatal exceptions, shall we?
Since there are a lot of moving parts, lots of production deployments,
and a short development-release cycle, it's a good idea to ease up on
any fatal errors that cause the plugin to cease any and all operations.

Switch to logging instead, no more exceptions. This will inherently
produce less strict APIs, but it's the way PHP and WordPress have always
been, so let's give everyone a break :)