Quick fix for edit post image bug
When saving an edited entry the post image is deleted, this is a quick
fix without tests for now.

And some automated tests to make sure the fix works.

Note: there's still an issue with our upload fields if they are the only
ones being uploaded, since the entry goes through a state of having the
upload fields empty. A rewrite of the extension (perhaps as a standalone
plugin like the rest?) is overdue.