Fix post meta shortcode parsing
`is_string` on the wrong thing here, we should be checking whether the
meta value is a string or not, all meta keys are strings. This bug
resulted in the following sort of warnings being thrown down the
pipeline:

"preg_match_all() expects parameter 2 to be string, array given"

Tests + fix should take care of this little bug.