Add form_id configuration property to \GV\Field
form_id is required for \GV\GF_Fields to be initialized from View
configuration dumps. The current dumps do not provide this form_id, thus
we need a mechanism by which it "appears" there.

Added a new filter (with back-compatibility), created a filter in our
mocks to add form_id from a View object, added tests + a couple of
fixes.

Also, made sure the field configuration is accessbile in overridden
__get method in \GV\GF_Field.