Fix creditcard appearing with no credit card field
If product fields are present, the creditcard field is automatically
added to the list of available fields. This is a bug. The creditcard
field is a standalone field (at least in 2.x) and should only show up
(and does show up correctly) if the field is actually added.

Change the group to "payment", preventing it from being added along with
all "product" fields when product is detected.

`GFCommon::has_product_field` and `GFCommon::has_creditcard_field`
seemingly attest to the logic of this decision and fix.