Meet the Legacy_Context mock manager
This allows us to quickly mock the legacy state using a basic
configuration with our new objects, which magically hydrates the needed
singletons and global state.

This will be used around the ported code where old filters, actions,
globals, etc. are still expected to be accessed by the thousands of
lines of custom code out there.

Some early samples of its usage are included. For example, the delete
links in the templates require the global view ID. The back link in the
single entry template requires it too. Mocking it now is effortless.