Fix speed issue due to wp_debug_backtrace_summary (#903)
Slow! :unamused:

To add a bit here, for future us: it was found that our logging actions took quite a bit of time to complete, specifically the debug logging actions, which are sprinkled generously around. We came to the conclusion that debug logs don't need backtraces, which are the primary slowdown here.