\GV\GF_Entry::from_entry performance shortcut
Gravity Forms does not cache entries in the object cache like WordPress
does with posts and whatnot. I should have checked to make sure it does,
but WordPress made me lazy since it's so easy to just rely on it.

We can now use ::from_entry where needed. Especially in the
\GV\Entry_Collection::fetch callbacks inside the \GV\GF_Form::entries
method.