Refactored oEmbed, with WordPress 4.8 support
WordPress 4.8 forces us to become our own oEmbed provider.

The Add Media / From URL entry will perform a call to the oEmbed REST
API which in turn tries to fetch the oEmbed data from a provider.

If no provider is given, a default fetcher kicks in, which tries to
detect the Post ID and render it. But our URLs are a superset of the
View ID, resulting in the preview being completely messed up.

Pending security and tests.