Fix issue with checkboxes in edit-entry
When saving a custom field, our code will load the value of the saved
field back into the entry. For later versions of GF the value for
checkboxes is an array. So the `gform_field_values` would contain a
really bad encoded string that fails to validate. Saving again kills
at least custom field file uploads.

This fix checks to see whether the value is an array to begin with and
pulls out the needed subvalue.

https://secure.helpscout.net/conversation/394406485/9835/?folderId=408730