Fix transient cache issues when invalidating cache
Make sure that \GravityView_Cache::delete clears the object cache as
well, otherwise a call to \GravityView_Cache::get will grab the result
from it even once we've deleted it via the database.

Delete it via the API to make sure the object cache does not serve us
with stale data.

Tests included.