Fix entry oEmbeds when using "default" permalink format
When passed "example.com/?gravityview=example&entry=123",
`url_to_postid()` would always return the front page ID, which wouldn't
be true, so the output would fail.

Now we check against whether the home page is the post ID returned from
`url_to_postid()`, and if it is, we assume this is wrong.

If the home page IS a View, then the only problem this causes is an
extra query, which should be cheap. Not a problem there.

Fixes #927