Fix DataTables entry linking (with DT patch coming)
The entry link functionality uses GravityView_API::directory_link(),
which includes a check for DOING_AJAX and $_POST['post_id']. This is
there for DataTables, but it never gets to be set because $post_id
isn't empty!

So we leave $post_id empty and it'll be calculated by
GravityView_API::directory_link() instead.