Finalize page_size widget
Tests. I don't really like the current solution as it's too high level, but
it does allow us to redefine the view safely for all the contexts that come after.
"Come after" is the problem. During redefinition there's absolutely no
context right now. We don't know where the View is being displayed, etc.

If we find ourselves needing more context than just the View, the widget it holds
and the GET query parameter, then we'll have to hook a little later. Other than that
this looks good to me.