Fix full field values being moved to the end of the output row
Let's say you have a Name field and you want to show the full name, first name, and last name.

The $allowed field ids are already in the correct order, with the correct field IDs: `["1", "1.3", "1.6"]`

$return uses the values from the entry, which only includes the input values (not the parent field ID). The $return array will be something like `[ 'id' => 123, 'form_id' => 456, '1.3' => "First", '1.6' => "Last" ]`.

After running through the $allowed filter, we get `[ '1.3' => "First", '1.6' => "Last" ]`. But notice that `1` is not there.

Then we loop through the $allowed fields by field ID and `1` is first in line. But it's not already set in $return, so it gets added to the end of the $return array: `[ '1.3' => "First", '1.6' => "Last", '1' => "First Last" ]`.

So the order of the field headers is correct, but any full field values would get moved to the end of the array in the CSV renderer.

The fix? I removed $return because none of the values were being used (they were being overwritten).