Hotfix revert: images were turning into partial links
Revert 9743f17ed50ec7b61f02a2e7805a44980eb97f1d

We can add this back in after making sure that it's working properly.

The reason we added it is so that it uses the GF-generated download URL with the hash. It's not a full URL, so I think pathinfo() isn't generating the extension properly, so it's not being recognized as an image.