More edit entry pricing calcultion fixes
A major problem has been discovered with Pricing field quantities in
calculation fields. These were being output as 0 during recalculation.
This was fixed in the previous commit.

Futher tests showed that the quantity was not being saved at all either.
This fixes saving of quantities.

Whew!

IE2, where are youuuu?