Fix sorting
This issue was likely introduced by the introduction of an empty `?sort` parameter in dec13f6e249361855047487369cedd97fdf2a1ee, as well as multisort.

The main issue is that `$skip_first` was being applied too aggressively; if the URL being passed only included one sort, like `?sort[20]=asc`, it was being skipped.

When a sort is provided but it's empty, `updateViewSorting()` returns an empty array, which breaks things.

@soulseekah Can you write some tests that play with this some more?

Fixes #1323