Disable drag & drop when modals are open
This happens all the time for me: I'm in Custom Content and I accidentally grab the modal while trying to copy text. Instead, it starts a drag-and-drop event.

Disabling drag & drop while the modal is open allows for text selection.

There's no good reason someone would want to drag and drop a field or widget while the modal is open.