Add back-compatibility for [score] field setting
I converted the score setting to a hidden field type in order to preserve the data if someone edits a View without migrating the values.

The field_options() method lacks state, so I couldn't simply check whether `score` was set and if so, migrate the value there.

I chose against using the class-gravityview-migrate.php class because looping through Views and updating settings is far riskier than on-demand JS updates.

On the template side, I check if there's a `score` setting and if so, set `choice_display` to "score".

@mrcasual Please review my JS