Fix Utils::get() not working for magic methods.
This didn't work:

```
$choice_display = \GV\Utils::get( $field, 'choice_display', $default_display );
```

because `$field->choice_display` is fetched by a magic method and isn't a defined property for the object.

The \GV\Field class does have an `__isset()` magic method that checks whether the property exists.

So let's use isset() instead of `property_exists()` to make things work as originally intended!