Fix field details state not being properly maintained
When opening fields, it should pull from the cookie and set whether or not to show field details. The container being used didn't include the element that was supposed to be targeted, so I had to go up a layer in the UI.

Not sure why this wasn't caught earlier.