Plug a big performance hole
Big performance issue here: every one of these fields being rendered led to the form being fetched fresh from the database:

- Custom Content
- Delete Link
- Duplicate Link
- Edit Entry Link
- Link to Single Entry

This is one of the reasons we were seeing queries increase linearly with the number of forms being added to a View.

I'm glad we found this. I'm sad it took so long.