Add a way to return bool from Request::is_view()
We were getting errors on the All Views screen because the Support Port wanted to know if we were on a View CPT page or not. It checked using `$request->is_view();`.

We didn't need a full instantiation of the View, we just needed a boolean response.

This shaves off a few hundredths of a second.

@mrcasual I'm not sure I like the structure of the parameter—'bool' looks ugly—but let me know what you think.