Complete updated merge tag support
When processing hundreds (or thousands!) of merge tags on a page with hundreds of fields, that can slow browsers down to a crawl. One customer had 5 minutes of render time dedicated to this.

So in the DOM, we declare merge tag support using `gv-merge-tag-support` instead of `merge-tag-support`. Then, when field settings dialogs are opened, we re-initialize the merge tags just-in-time.