Parse the current global $post to detect Views
This is a big change, but I believe it's a Good Thing™. `gravityview()->views->get()` _should_ return all the Views available, not only if we're inside a CPT.

This modifies `get()` to parse the global `$post` object to also check for embedded shortcodes and blocks. The perfect lack of reliance on the global state is somewhat removed, but it's adding something we needed for GravityKit/Social-Sharing#28 and, I think, a lot of other page builder issues.

@mrcasual, please review this!