Have View Settings tabs remember what tab was open
It's annoying to be working on a View and configuring some setting and having to re-open that setting over and over again. This has been something I've been wanting to do forever.

When talking with Blaz from @jetsloth, I realized the key is that the cookie should expire with the session—I had thought it would have had to persist.

I believe this change will make everyone's lives easier.