Remove `embed_js_in_footer`
This setting—about just the custom JS code—would have instead changed the place where the `gravityview-fe-view` script was being enqueued.

This change wouldn't really help anything, so I'm removing it. As far as I know, no one has ever asked to enqueue the script in the header.

Sorry for the back-and-forth on this!