Limit the use of parameters to generate a cache key (#2089)
We use the View settings object to generate a unique
cache key that's used to query the database for entries
and cache the results. While the query can be the same,
the settings object can change during the request, so
a duplicate query is executed.

This limits those parameters to a "cache_atts" property on
the settings object, which can be used to specifically control
how the cache key is generated.