Fallback to View type selection when data-templateid is missing
https://github.com/GravityKit/GravityView/commit/0fc033b145063cd6ab16625e13031ab79f54bf59
requires a templateid data attribute, which is not available when Multiple Forms is enabled.
While this attribute will be set on the Add Field button in the next Multiple Forms release,
we can use the View type select element under each layout to retrieve the template ID as a
fallback.