Fix situation where array contains null value
I got a PHP 8.2 warning about running rawurldecode() when the first parameter is null.

The $get variable was an array:

array:3 [▼
  "post" => "514"
  "action" => "edit"
  "id" => null
]

that was being passed to gv_map_deep() in class-search-widget.php:

```
$get = gv_map_deep( $get, 'rawurldecode' );
```

I figured this should be fixed inside gv_map_deep() so it's solved for all other uses.

Related to #1999