Fixes RAND sorting issues
Resolves #2211

The error was being caused by the code expecting an array and we were just passing a string for the sort direction. But that code shouldn't be running if there's RAND anyway; sorting doesn't matter in that situation.

So I added a check to determine whether randomness is being applied. If so, skip the multi-sort code anyway.