Do not extract duplicate Views from post content
In the context of embedded Views, we add `?gvid=X` to the URL
based on the number of Views found. Using Elementor, for example,
a single View embed would be counted twice, adding the unencessary
query parameter. This also had a negative side-effect when opening
a Single Entry entry layout where one of the fields would embed
another View.