Include our blocks in the Getting Started page
I noticed we were still encouraging people to use shortcodes in the block editor. This updates to use the blocks we have.

It also uses some new styling. I think it looks nice, and it's temporary until we get the proper onboarding redesign implemented.

I recorded a short video.

I also translated a bunch of strings, which is okay now because we have ChatGPT to help with it!