Return expected response format (with no entries)
I don't understand any scenario where you would want to have an error state when there are no entries. It seems like looping through an entry set of zero and returning the correct headers is the expected state. So I think that this code was a bug to begin with.

I've looked throughout our codebases and we don't have any expectation of the `gravityview-no-entries` error code being returned. Also, I searched for the string "No Entries found." and found no instances of that as well.

This seems to be the only place in our code where this response is used. I believe it is safe to remove it.

Resolves #2445