Don't show `show_in_shortcode` shortcode setting
The setting is used in the old GravityView_Admin_Add_Shortcode class, but it's also used by the Elementor widget to determine what settings should be exposed.

It's not actually possible to override this setting using the shortcode; that affects the visibility of entries and could result in entries being revealed that weren't meant to be.

I'm removing this for the GV Elementor widget.